refactor(backend): relocate Cooling enum to Anela.Heblo.Domain.Shared
Moves the Cooling enum (None/L1/L2) from Anela.Heblo.Domain.Features.Catalog
to Anela.Heblo.Domain.Shared so that ShoptetOrders, Logistics, Packaging,
CarrierCooling, the ShoptetApi adapter, and the Flexi adapter no longer take
a compile-time dependency on the Catalog module solely to reach a
temperature-chain shipping classification. The enum's CLR identity, integer
values, EF Core string-conversion persistence, and OpenAPI surface are
unchanged.

Also adds missing ICarrierCoolingRepository and IGiftSettingRepository mocks
to ShoptetIntegrationTestFixture to fix pre-existing integration test failures.
