+++ backend/src/Anela.Heblo.Application/Features/CarrierCooling/Contracts/CarrierCoolingRowDto.cs
@@ -1,5 +1,5 @@
-using Anela.Heblo.Domain.Features.Catalog;
 using Anela.Heblo.Domain.Features.Logistics;
+using Anela.Heblo.Domain.Shared;
 
 namespace Anela.Heblo.Application.Features.CarrierCooling.Contracts;
