+++ backend/src/Anela.Heblo.Domain/Features/Catalog/Attributes/CatalogAttributes.cs
@@ -1,4 +1,5 @@
 using Anela.Heblo.Domain.Features.Catalog;
+using Anela.Heblo.Domain.Shared;
 
 namespace Anela.Heblo.Domain.Features.Catalog.Attributes;
