+++ backend/src/Anela.Heblo.Domain/Features/Logistics/CarrierCoolingSetting.cs
@@ -1,4 +1,4 @@
-using Anela.Heblo.Domain.Features.Catalog;
+using Anela.Heblo.Domain.Shared;
 
 namespace Anela.Heblo.Domain.Features.Logistics;
