+++ backend/test/Anela.Heblo.Tests/Controllers/CarrierCoolingControllerTests.cs
@@ -4,8 +4,8 @@ using System.Text.Json;
 using System.Text.Json.Serialization;
 using Anela.Heblo.Application.Features.CarrierCooling.UseCases.GetCarrierCoolingMatrix;
 using Anela.Heblo.Application.Features.CarrierCooling.UseCases.SetCarrierCooling;
-using Anela.Heblo.Domain.Features.Catalog;
 using Anela.Heblo.Domain.Features.Logistics;
+using Anela.Heblo.Domain.Shared;
 using Anela.Heblo.Tests.Common;
 using FluentAssertions;
 using Xunit;
